FAC-29 chore: refactored file structure and added verify script #52
