+++ src/modules/questionnaires/dto/requests/create-version-request.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsObject, IsNotEmpty } from 'class-validator';
-import type { QuestionnaireSchemaSnapshot } from '../../questionnaire.types';
+import type { QuestionnaireSchemaSnapshot } from '../../lib/questionnaire.types';
 
 export class CreateVersionRequest {
   @ApiProperty()
