+++ src/modules/questionnaires/services/scoring.service.spec.ts
@@ -4,7 +4,7 @@ import {
   QuestionnaireSchemaSnapshot,
   QuestionnaireType,
   QuestionType,
-} from '../questionnaire.types';
+} from '../lib/questionnaire.types';
 
 describe('ScoringService', () => {
   let service: ScoringService;
